support build of images on multiple archs (#23)
Buildah action takes `archs` argument, which labels the built container with specified architecture. The build itself needs to be done in qemu. This commit introduces new optional input for the build-and-push-action named `archs` where user can specify one or more architectures to build images on. The default arch is amd64. The qemu tooling is installed in separate step.
push-to-registry is bumped from version 2.2 to version 2 - currently pointing to 2.7.